moved get_first_purchase_offer_banner_fragment to Discounts utils.py
